text-processing/src/requirements.txt
torch==1.7.1
sentence-transformers
spacy
bert-score
click
transformers
sentencepiece
pandas
flair
stanford_openie
absl-py
six
rouge-score